fix: wip cli
