+++ terminator-cli/src/workflow_exec.rs
@@ -1,96 +1,6 @@
+use anyhow::Context;
 use crate::mcp_client;
-use anyhow::{Context, Result};
-use std::process::{Command, Stdio};
-use crate::cli::{InputType, McpRunArgs, McpCommands};
-
-pub fn run_command(program: &str, args: &[&str]) -> Result<(), Box<dyn std::error::Error>> {
-    let output = Command::new(program)
-        .args(args)
-        .stdout(Stdio::piped())
-        .stderr(Stdio::piped())
-        .output()?;
-
-    if !output.status.success() {
-        let stderr = String::from_utf8_lossy(&output.stderr);
-        return Err(format!(
-            "Command failed: {} {}\nError: {}",
-            program,
-            args.join(" "),
-            stderr
-        )
-        .into());
-    }
-
-    Ok(())
-}
-
-pub fn handle_mcp_command(cmd: McpCommands) {
-    let transport = match cmd {
-        McpCommands::Chat(ref args) => parse_transport(args.url.clone(), args.command.clone()),
-        McpCommands::AiChat(ref args) => parse_transport(args.url.clone(), args.command.clone()),
-        McpCommands::Exec(ref args) => parse_transport(args.url.clone(), args.command.clone()),
-        McpCommands::Run(ref args) => parse_transport(args.url.clone(), args.command.clone()),
-    };
-
-    let rt = tokio::runtime::Runtime::new().expect("Failed to create Tokio runtime");
-
-    let result = rt.block_on(async {
-        match cmd {
-            McpCommands::Chat(_) => mcp_client::interactive_chat(transport).await,
-            McpCommands::AiChat(_) => mcp_client::natural_language_chat(transport).await,
-            McpCommands::Exec(args) => {
-                mcp_client::execute_command(transport, args.tool, args.args).await
-            }
-            McpCommands::Run(args) => run_workflow(transport, args).await,
-        }
-    });
-
-    if let Err(e) = result {
-        eprintln!("❌ MCP command error: {e}");
-        std::process::exit(1);
-    }
-}
-
-pub fn parse_transport(url: Option<String>, command: Option<String>) -> mcp_client::Transport {
-    if let Some(url) = url {
-        mcp_client::Transport::Http(url)
-    } else if let Some(command) = command {
-        let parts = parse_command(&command);
-        mcp_client::Transport::Stdio(parts)
-    } else {
-        // Default to spawning local MCP agent via npx for convenience
-        let default_cmd = "npx -y terminator-mcp-agent@latest";
-        println!("ℹ️  No --url or --command specified. Falling back to '{default_cmd}'");
-        let parts = parse_command(default_cmd);
-        mcp_client::Transport::Stdio(parts)
-    }
-}
-
-pub fn parse_command(command: &str) -> Vec<String> {
-    // Simple command parsing - splits by spaces but respects quotes
-    let mut parts = Vec::new();
-    let mut current = String::new();
-    let mut in_quotes = false;
-
-    for c in command.chars() {
-        match c {
-            '"' => in_quotes = !in_quotes,
-            ' ' if !in_quotes => {
-                if !current.is_empty() {
-                    parts.push(current.clone());
-                    current.clear();
-                }
-            }
-            _ => current.push(c),
-        }
-    }
-
-    if !current.is_empty() {
-        parts.push(current);
-    }
-
-    parts
-}
+use crate::cli::{InputType, McpRunArgs};
 
 pub async fn run_workflow(transport: mcp_client::Transport, args: McpRunArgs) -> anyhow::Result<()> {
     use tracing::info;
